Update state_machine.c
bugfix float to double for stooklijn-correction-parameters

+++ state_machine.c
@@ -5,8 +5,8 @@
             static States newstate = Idle;
             static float water_hyst_pos = 4.0; //still to be replaced by LG modbus state if we find how to connect this
             static float water_hyst_neg = -4.0; //still to be replaced by LG modbus state if we find how to connect this
-            static float max_stooklijn_correction_pos = 3.0;
-            static float max_stooklijn_correction_neg = -5.0;
+            static double max_stooklijn_correction_pos = 3.0;
+            static double max_stooklijn_correction_neg = -5.0;
             static uint32_t timer = 0;
             static uint32_t statechange = 0; //timer value upon previous state change
             static uint32_t compressortime = 0; //timer value on last compressor start
